add routers and global config manager class
