+++ FastAPI_server/globals/globals_instance.py
@@ -1,0 +1,3 @@
+from ModelForge.FastAPI_server.globals.globals import GlobalSettings
+
+global_manager = GlobalSettings.get_instance()
