+++ Frontend/src/services/api.js
@@ -1,5 +1,5 @@
 export const config = {
-    baseURL: '/api',
+    baseURL: 'http://localhost:8000',
     headers: {
         'Content-Type': 'application/json',
     },
